Add files via upload
This is a python file for the classic number guessing game. I'm tripping over an issue where I want to look at the previous guess and determine if it was closer or further away than the current guess. I think my issue is on lines 57 and 87, where, at the end of the while loop, I set the current guess as the previous guess and start over, but at the top, the guess is set to the function. I need a way to delineate these two. plz advise.